Moving reader code for kinemage into its extensions directory
