+++ src/extensions/kinemage/behavior.ts
@@ -18,7 +18,7 @@ import { PluginContext } from '../../mol-plugin/context';
 import { DefaultQueryRuntimeTable } from '../../mol-script/runtime/query/compiler';
 import { StateTransforms } from '../../mol-plugin-state/transforms';
 import { shapePointsFromKin, shapeLinesFromKin, shapeMeshFromKin, shapeSpheresFromKin } from '../../mol-model-formats/shape/kin';
-import { Kinemage } from '../../mol-io/reader/kin/schema';
+import { Kinemage } from './reader/schema';
 import { DataFormatProvider } from '../../mol-plugin-state/formats/provider';
 import { Camera } from '../../mol-canvas3d/camera';
 import { PluginCommands } from '../../mol-plugin/commands';
