+++ src/extensions/kinemage/reader/schema.ts
